Add files via upload
We got our data from fellow classmates and friends. For the student’s address, those that weren’t 
comfortable with sharing their address received a similar address of housing on campus. We 
asked for the courses they are enrolled in and went to UMD’s website Testudo to get more 
information about the course and instructor. We also got information about the department on 
Testudo.